+++ client/src/components/app/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import '../../scss/style.module.scss';
 import HomePage from '../pages/home';
+import '../../scss/style.module.scss';
 
 function App(): JSX.Element {
   return <HomePage />;
