+++ client/src/components/blocks/collectionMovies/CollectionMovies.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
 import React from 'react';
 import Cards from './cards';
 import SideBar from './sideBar';
