+++ client/src/components/blocks/collectionMovies/sideBar/menu/index.ts
@@ -1,3 +1,5 @@
 import Menu from './Menu';
 
+export type { IMenuItem } from './menuItem/models';
+
 export default Menu;
